+++ src/components/question-types/EssayAnswerTextarea.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Textarea } from "@/components/ui/textarea";
+
+const EssayAnswerTextarea = () => (
+  <Textarea placeholder="긴 서술형 답변" className="min-h-[100px]" />
+);
+
+export default EssayAnswerTextarea;
