+++ src/components/question-types/ShortAnswerInput.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Input } from "@/components/ui/input";
+
+const ShortAnswerInput = () => (
+  <Input placeholder="짧은 주관식 답변" className="w-full h-10" />
+);
+
+export default ShortAnswerInput;
